+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Added Persian numerals option to `format` function (third parameter)
+
 ## 0.6.1
 - Fixed `format` token replacement conflict bug